# lab_wizard/wizard/frontend/src/lib/api/openapi.json
{
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
   "_ForceReleaseClaimRequest": {
    "properties": {
     "unit": {
      "title": "Unit",
      "type": "string"
     },
     "url": {
      "title": "Url",
      "type": "string"
     }
    },
    "required": [
     "url",
     "unit"
    ],
    "title": "_ForceReleaseClaimRequest",
    "type": "object"
   },
...
   "_ReleaseRequest": {
    "properties": {
     "path": {
      "title": "Path",
      "type": "string"
     },
     "url": {
      "title": "Url",
      "type": "string"
     }
    },
    "required": [
     "url",
     "path"
    ],
    "title": "_ReleaseRequest",
    "type": "object"
   },
   "_RemoteDiscoverRequest": {
    "properties": {
     "action": {
      "title": "Action",
      "type": "string"
     },
     "config_dir": {
      "title": "Config Dir",
      "type": "string"
     },
     "params": {
      "additionalProperties": true,
      "title": "Params",
      "type": "object"
     },
     "parent_chain": {
      "items": {},
      "title": "Parent Chain",
      "type": "array"
     },
     "type": {
      "title": "Type",
      "type": "string"
     }
    },
    "required": [
     "config_dir",
     "type",
     "action"
    ],
    "title": "_RemoteDiscoverRequest",
    "type": "object"
   },
   "_RemoteEditRequest": {
    "properties": {
     "config_dir": {
      "title": "Config Dir",
      "type": "string"
     },
     "operation": {
      "title": "Operation",
      "type": "string"
     },
     "payload": {
      "additionalProperties": true,
      "title": "Payload",
      "type": "object"
     }
    },
    "required": [
     "config_dir",
     "operation"
    ],
    "title": "_RemoteEditRequest",
    "type": "object"
   },
   "_RemoteServerRequest": {
    "properties": {
     "name": {
      "title": "Name",
      "type": "string"
     },
     "url": {
      "title": "Url",
      "type": "string"
     }
    },
    "required": [
     "name",
     "url"
    ],
    "title": "_RemoteServerRequest",
    "type": "object"
   },
   "_RemoteTestRequest": {
    "properties": {
     "url": {
      "title": "Url",
      "type": "string"
     }
    },
    "required": [
     "url"
    ],
    "title": "_RemoteTestRequest",
    "type": "object"
   },
   "_RemoteTreeRequest": {
    "properties": {
     "config_dir": {
      "title": "Config Dir",
      "type": "string"
     }
    },
    "required": [
     "config_dir"
    ],
    "title": "_RemoteTreeRequest",
    "type": "object"
   },
...
   "_SavePermissionsRequest": {
    "properties": {
     "permissions": {
      "additionalProperties": true,
      "title": "Permissions",
      "type": "object"
     }
    },
    "title": "_SavePermissionsRequest",
    "type": "object"
   },
...
   "_ServerBindRequest": {
    "properties": {
     "bind": {
      "title": "Bind",
      "type": "string"
     }
    },
    "required": [
     "bind"
    ],
    "title": "_ServerBindRequest",
    "type": "object"
   },
   "_ServerStartRequest": {
    "properties": {
     "detached": {
      "default": false,
      "title": "Detached",
      "type": "boolean"
     }
    },
    "title": "_ServerStartRequest",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "FastAPI",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/local-servers/claims/force-release": {
   "post": {
    "description": "End every claim touching ``unit`` on one server.\n\nFor a run that is stuck, or whose client disappeared. The server resets the\nreleased instruments to baseline before anyone else can claim them.",
    "operationId": "api_force_release_claim_api_local_servers_claims_force_release_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_ForceReleaseClaimRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Force Release Claim"
   }
  },
  "/api/local-servers/release": {
   "post": {
    "description": "Ask a server to disconnect and evict one root.\n\nHands a rack back without stopping the whole server, which is what a user\nwants when a local project needs the bus. The path stays servable \u2014 the next\ncall through the server reopens it.",
    "operationId": "api_release_hardware_api_local_servers_release_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_ReleaseRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Release Hardware"
   }
  },
...
  "/api/permissions": {
   "get": {
    "description": "Return the local instrument tree + permission vocabulary + current rules.\n\n``tree`` mirrors manage-instruments; ``instruments`` carries the state keys\nand methods the rule builder offers per node; ``permissions`` is the current\n``permissions:`` block from server.yaml.",
    "operationId": "api_get_permissions_api_permissions_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "Api Get Permissions"
   },
   "put": {
    "description": "Validate and persist the ``permissions:`` block to server.yaml.",
    "operationId": "api_save_permissions_api_permissions_put",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_SavePermissionsRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Save Permissions"
   }
  },
...
  "/api/remote-servers": {
   "get": {
    "description": "Return the registered remote servers (the client address book).",
    "operationId": "api_get_remote_servers_api_remote_servers_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "Api Get Remote Servers"
   },
   "post": {
    "description": "Add (or update by name) a remote server.",
    "operationId": "api_add_remote_server_api_remote_servers_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_RemoteServerRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Add Remote Server"
   }
  },
  "/api/remote-servers/test": {
   "post": {
    "description": "Live-test a remote server URL and return its attributes (never errors).",
    "operationId": "api_test_remote_server_api_remote_servers_test_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_RemoteTestRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Test Remote Server"
   }
  },
  "/api/remote-servers/{name}": {
   "delete": {
    "description": "Remove a remote server by name.",
    "operationId": "api_remove_remote_server_api_remote_servers__name__delete",
    "parameters": [
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Remove Remote Server"
   }
  },
  "/api/remote-tree": {
   "post": {
    "description": "Tree, schema and recent activity of another workspace's server.\n\nThe schema comes from that server, not this build: it decides which\ninstrument types exist and what fields they take.",
    "operationId": "api_remote_tree_api_remote_tree_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_RemoteTreeRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Remote Tree"
   }
  },
  "/api/remote-tree/discover": {
   "post": {
    "description": "Run a discovery scan on another workspace's server.\n\nThe scan has to happen where the hardware is. Without it, adding an\ninstrument to another workspace would mean knowing its bus address by heart.",
    "operationId": "api_remote_discover_api_remote_tree_discover_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_RemoteDiscoverRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Remote Discover"
   }
  },
  "/api/remote-tree/edit": {
   "post": {
    "description": "Add, remove, or reset an instrument on another workspace's server.\n\nThe server enforces both rules that matter \u2014 same-machine only, and not\nwhile the rack is open \u2014 so a refusal here is its answer, not ours.",
    "operationId": "api_remote_tree_edit_api_remote_tree_edit_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_RemoteEditRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Remote Tree Edit"
   }
  },
  "/api/remote-tree/events": {
   "post": {
    "description": "Recent notable events recorded by that server.",
    "operationId": "api_remote_events_api_remote_tree_events_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_RemoteTreeRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Remote Events"
   }
  },
  "/api/server/bind": {
   "put": {
    "description": "Persist this workstation's server bind address.",
    "operationId": "api_server_set_bind_api_server_bind_put",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_ServerBindRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Server Set Bind"
   }
  },
...
  "/api/server/restart": {
   "post": {
    "description": "Restart the server \u2014 used to apply edited permission rules.",
    "operationId": "api_server_restart_api_server_restart_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_ServerStartRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Server Restart"
   }
  },
  "/api/server/start": {
   "post": {
    "description": "Start the instrument server (managed child or detached daemon).",
    "operationId": "api_server_start_api_server_start_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_ServerStartRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Server Start"
   }
  },
...
  "/api/server/suggest-port": {
   "get": {
    "description": "Return a tcp://host:port bind on a currently-free port (not persisted).\n\n``prefer_default=true`` offers the standard/existing port first when free \u2014\nused for the initial suggestion before the server is configured.",
    "operationId": "api_server_suggest_port_api_server_suggest_port_get",
    "parameters": [
     {
      "in": "query",
      "name": "prefer_default",
      "required": false,
      "schema": {
       "default": false,
       "title": "Prefer Default",
       "type": "boolean"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Server Suggest Port"
   }
  },
...
 }
}